moved completely to floats for amts, added display option for measurements for the main display, and added Measurements class to do calculations of multiple types

+++ ConversionsTake2/src/main/java/Unit.java
@@ -12,7 +12,7 @@ public enum Unit implements Comparable<Unit> {
         this.multiplier = multiplier;
     }
 
-    private float asBase(int amt) {
+    private float asBase(float amt) {
         return amt * multiplier;
     }
